+++ Tests/Binacle.Net.Lib.UnitTests/BinacleStrategiesFixture.cs
@@ -11,9 +11,9 @@ namespace Binacle.Net.Lib.Tests
         {
             this.Bins = new List<Item>()
             {
-                new Item("Small", new Dimensions<ushort>(8,45,62)),
-                new Item("Medium", new Dimensions<ushort>(17,45,62)),
-                new Item("Large", new Dimensions<ushort>(36,45,62))
+                new Item("Small", new Dimensions<int>(8,45,62)),
+                new Item("Medium", new Dimensions<int>(17,45,62)),
+                new Item("Large", new Dimensions<int>(36,45,62))
             };
 
         }
